misc/HtmlTestSuite/results/test-0145.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<link href="mailto:dholbert@mozilla.com" rel="author" title="Daniel Holbert"/>
<link href="mailto:tlin@mozilla.com" rel="author" title="Ting-Yu Lin"/>
<link href="https://www.mozilla.org/" rel="author" title="Mozilla"/>
<link href="https://drafts.csswg.org/css-sizing-3/#intrinsic-sizes" rel="help"/>
<link href="https://bugzilla.mozilla.org/show_bug.cgi?id=1638937" rel="help"/>

<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>

<style>
.flexbox {
  display: flex;
  width: 600px;
  border: 2px solid black;
}
svg {
  background: teal;
  flex: none;
}
</style>

</head><body onload="checkLayout('svg')">
  No intrinsic attributes:
  <div class="flexbox">
    <svg xmlns="http://www.w3.org/2000/svg" data-expected-client-height="150" data-expected-client-width="300"/>
  </div>

  viewBox and height:
  <div class="flexbox">
    <svg xmlns="http://www.w3.org/2000/svg" data-expected-client-height="40" data-expected-client-width="160" height="40" viewbox="0 0 4 1"/>
  </div>

  viewBox and width:
  <div class="flexbox">
    <svg xmlns="http://www.w3.org/2000/svg" data-expected-client-height="10" data-expected-client-width="40" viewbox="0 0 4 1" width="40"/>
  </div>

  viewBox, width, height:
  <div class="flexbox">
    <svg xmlns="http://www.w3.org/2000/svg" data-expected-client-height="40" data-expected-client-width="40" height="40" viewbox="0 0 4 1" width="40"/>
  </div>

  Just viewBox:
  <div class="flexbox">
    <svg xmlns="http://www.w3.org/2000/svg" data-expected-client-height="150" data-expected-client-width="600" viewbox="0 0 4 1"/>
  </div>
</body>
</html>
